+++ Source/SSCore/src/main/java/com/realtech/socialsurvey/core/entities/UserRankingPastYearsRegion.java
@@ -66,7 +66,7 @@ public class UserRankingPastYearsRegion
     private int totalReviews;
     
     @Column(name = "average_rating")
-    private int averageRating;
+    private float averageRating;
     
     @Column(name = "is_eligible")
     private	int isEligible;
@@ -231,15 +231,13 @@ public class UserRankingPastYearsRegion
         this.totalReviews = totalReviews;
     }
 
-    public int getAverageRating()
-    {
-        return averageRating;
-    }
+	public float getAverageRating() {
+		return averageRating;
+	}
 
-    public void setAverageRating( int averageRating )
-    {
-        this.averageRating = averageRating;
-    }
+	public void setAverageRating(float averageRating) {
+		this.averageRating = averageRating;
+	}
 
 	public int getIsEligible() {
 		return isEligible;
