+++ Source/SSCore/src/main/java/com/realtech/socialsurvey/core/entities/UserRankingThisYearRegion.java
@@ -68,7 +68,7 @@ public class UserRankingThisYearRegion
     private int totalReviews;
     
     @Column(name = "average_rating")
-    private int averageRating;
+    private float averageRating;
     
     @Column(name = "is_eligible")
     private	int isEligible;
@@ -82,11 +82,11 @@ public class UserRankingThisYearRegion
     @Transient
     private String profileImageUrlThumbnail;
     
-	public int getAverageRating() {
+	public float getAverageRating() {
 		return averageRating;
 	}
 
-	public void setAverageRating(int averageRating) {
+	public void setAverageRating(float averageRating) {
 		this.averageRating = averageRating;
 	}
